Add basic nuke build
