+++ samples/HexControl.Samples.WinUI/Assets/LockScreenLogo.scale-200.png
