+++ samples/HexControl.Samples.WinUI/Assets/SplashScreen.scale-200.png
