+++ samples/HexControl.Samples.WinUI/Assets/Square44x44Logo.scale-200.png
