+++ samples/HexControl.Samples.WinUI/Assets/Square44x44Logo.targetsize-24_altform-unplated.png
