+++ samples/HexControl.Samples.WinUI/Assets/StoreLogo.png
