--- samples/HexControl.Samples.WinUI/HexControl.Samples.WinUI/App.xaml.cs
@@ -1,34 +1,0 @@
-﻿using Microsoft.UI.Xaml;
-
-// To learn more about WinUI, the WinUI project structure,
-// and more about our project templates, see: http://aka.ms/winui-project-info.
-
-namespace HexControl.Samples.WinUI;
-
-/// <summary>
-///     Provides application-specific behavior to supplement the default Application class.
-/// </summary>
-public partial class App : Application
-{
-    private Window m_window;
-
-    /// <summary>
-    ///     Initializes the singleton application object.  This is the first line of authored code
-    ///     executed, and as such is the logical equivalent of main() or WinMain().
-    /// </summary>
-    public App()
-    {
-        InitializeComponent();
-    }
-
-    /// <summary>
-    ///     Invoked when the application is launched normally by the end user.  Other entry points
-    ///     will be used such as when the application is launched to open a specific file.
-    /// </summary>
-    /// <param name="args">Details about the launch request and process.</param>
-    protected override void OnLaunched(LaunchActivatedEventArgs args)
-    {
-        m_window = new MainWindow();
-        m_window.Activate();
-    }
-}
