--- samples/HexControl.Samples.WinUI/HexControl.Samples.WinUI/app.manifest
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<assembly manifestVersion="1.0" xmlns="urn:schemas-microsoft-com:asm.v1">
-	<assemblyIdentity version="1.0.0.0" name="HexControl.Samples.WinUI.app" />
-
-	<application xmlns="urn:schemas-microsoft-com:asm.v3">
-		<windowsSettings>
-			<dpiAware xmlns="http://schemas.microsoft.com/SMI/2005/WindowsSettings">true/PM</dpiAware>
-			<dpiAwareness xmlns="http://schemas.microsoft.com/SMI/2016/WindowsSettings">PerMonitorV2, PerMonitor</dpiAwareness>
-		</windowsSettings>
-	</application>
-</assembly>
