+++ samples/HexControl.Samples.WinUI/MainWindow.xaml
