+++ samples/HexControl.Samples.WinUI/MainWindow.xaml.cs
